model, controller, and routes synced up - handlebars synced up with mvc - bootstrap added - css styling added
