+++ models/Headline.js
@@ -2,8 +2,11 @@ const mongoose = require("mongoose");
 
 const Schema = mongoose.Schema;
 
+const request = require('request');
+const cheerio = require('cheerio');
+
 const headlineSchema = new Schema({
-    
+
     date: {
         type: String,
         required: true
@@ -13,7 +16,7 @@ const headlineSchema = new Schema({
         type: String,
         required: true
     },
-    
+
     author: {
         type: String,
         required: true
@@ -32,9 +35,71 @@ const headlineSchema = new Schema({
     },
 });
 
+
+headlineSchema.statics.scrape = function (callback) {
+
+    const hostName = "https://www.naturalnews.com/";
+
+    let initialScrapeArray = []
+
+    request(hostName, function (error, response, html) {
+
+        if (error) console.log(error);
+
+        const $ = cheerio.load(html);
+
+        $("div.IFB").each(function (i, element) {
+            // Save the text and href of each link enclosed in the current element
+            let date = $(element).children("div.IFA").children("time.entry-time").text();
+            let title = $(element).children("div.IFH").children("a").text();
+            let author = $(element).children("div.IFA").children("span.entry-author").children("a").children().text();
+            let imagePath = $(element).children("div.IFP").children("a").children().attr("src");
+            let description = $(element).children("div.IFD").find("p").text();
+            let link = $(element).children("div.IFH").children("a").attr("href");
+            let imageUrl = hostName + imagePath;
+
+            let articleData = {
+                date: date,
+                title: title,
+                author: author,
+                imageUrl: imageUrl,
+                description: description,
+                link: link
+            };
+
+
+            if (date && title && author && imageUrl && description && link) {
+
+                initialScrapeArray.push(articleData);
+
+                Headline.create(articleData, function (err, inserted) {
+                    if (err) {
+
+                        console.log(err);
+                    } else {
+
+                        console.log(inserted + 'HERE --------------------*************');
+                    }
+                });
+            }
+        });
+        callback(initialScrapeArray);
+
+    });
+
+};
+
+headlineSchema.statics.dbQuery = function (callback) {
+    Headline.find({}, function(err, news){
+        if (err) {
+            console.log(err);
+        } else {
+            callback(news);
+        }
+    });
+};
+
 const Headline = mongoose.model("Headline", headlineSchema);
 
+
 module.exports = Headline;
-
-
-
