+++ routes/view/html-routes.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const router = express.Router();
+
+const fetch = require('../../controllers/fetch.js');
+
+router.get('/', fetch.renderHeadlines);
+
+module.exports = router;
